+++ noodles-cram/src/io/reader/header/container/header.rs
@@ -5,14 +5,14 @@ use flate2::CrcReader;
 
 use crate::io::reader::num::{read_itf8, read_ltf8};
 
-pub(super) fn read_header<R>(reader: &mut R) -> io::Result<usize>
+pub(super) fn read_header<R>(reader: &mut R) -> io::Result<u64>
 where
     R: Read,
 {
     let mut crc_reader = CrcReader::new(reader);
 
     let length = crc_reader.read_i32::<LittleEndian>().and_then(|n| {
-        usize::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
+        u64::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
     })?;
 
     // reference sequence ID
